Use x86_64_*_operand in LEA splitters
Similar to the previous fix, these patterns need to use a constraint
that will reject a SYMBOL_REF expression in -m64 -fPIC.  That is not
how these patterns were handled in upstream -mx32 support.  But the
situation is substantially different there and possibly this is not
the relevant change there because Pmode is SImode for -mx32 while
it's still DImode for x86_64-nacl-gcc -m64.  (That said, I don't
actually understand what RTL transformations happen differently in
this case upstream under -mx32 to prevent it from exhibiting this
bug--which it does not--without this constraint change.)

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3598
TEST= toolchain trybots
R=khim@chromium.org

Review URL: https://codereview.chromium.org/21318002