Fix interlibrary dependencies in target shared library builds
The core change here is just the libtool.m4 change.  The purpose of
that is to teach libtool that *-*-nacl* is like *-*-linux* as far as
the semantics of dependencies between shared libraries.  The lack of
this meant that libstdc++.so did not have proper dependencies on
libc.so and libm.so.  Fixes to the linker mean that links using
--as-needed that happened to draw in libm.so as a dependency of the
final nexe no longer do, leading to undefined symbol errors for
libstdc++.so's references to libm.so symbols.

All the */configure changes are just due to regenerating with
Autoconf.  The status quo ante included a hodge-podge of hand edits
to the generated configure files as well as some changes to .m4
files that had then affected some configure files but not all
because the person doing the change regenerated some affected files
and not others.  The config/lib-link.m4 change is necessary to
preserve a change that was made with hand edits to configure files
but should have been made in that source file instead.

BUG= https://code.google.com/p/nativeclient/issues/detail?id=3636
TEST= local build and hand examination
R=sehr@chromium.org, mseaborn@chromium.org

Review URL: https://codereview.chromium.org/65643012